requirements.txt
discord.py
random
dotenv
google.generativeai